Delete BranchController, Delete RoleController, Adjust BaseResponseObj, Add GetPagingBranch in Model Branch, adjust ForeignKey to BigInt

+++ database/migrations/2024_08_15_090745_create_branches_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('branch_code', 8)->unique();
             $table->string('branch_name');
-            $table->integer('tenant_id');
+            $table->bigInteger('tenant_id');
             $table->string('address');
             $table->string('city');
             $table->string('provice');
